% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{anes2008_complete_cases}
\alias{anes2008_complete_cases}
\title{Data from the 2008 American National Election Study (ANES)}
\format{
A data frame with 11 variables: \code{religion}, \code{married},
\code{education}, \code{income}, \code{time}, \code{vote}, \code{jobs_r},
\code{imp_enviro}, \code{female}, \code{white,} and \code{age}.
}
\usage{
anes2008_complete_cases
}
\description{
This is a subset of the complete cases on 11 variables to use for evaluating
multiple imputation algorithms. The missingness is introduced by a known
mechanism in a separate step.
}
\details{
Please refer to Kropko, Jonathan; Goodrich, Ben; Gelman, Andrew; Hill, Jennifer, 2014,
"Replication data for: Multiple Imputation for Continuous and Categorical Data: Comparing Joint Multivariate Normal and Conditional Approaches",
https://doi.org/10.7910/DVN/24672, Harvard Dataverse, V3, UNF:5:QuxE8nFhbW2JZT+OW9WzWw== \link{fileUNF}
for more information on the original data.
}
\keyword{datasets}
