% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_MIBench_results.R
\name{get_MIBench_results}
\alias{get_MIBench_results}
\title{Get the results from the output of repeat_MIBench_experiment}
\usage{
get_MIBench_results(obj)
}
\arguments{
\item{obj}{The output of a run of \code{repeat_MIBench_experiment}}
}
\value{
A list of class "MIbench_imputations"
}
\description{
Get the results from the output of repeat_MIBench_experiment
}
\examples{
df <- cars
regression <- lm(log(dist)~speed, data = df)
# Specifiying no scenario to simulate at the mean of speed.
simloglm(regression)
# Explicitily specifying a scenario.
simloglm(regression, scenario = list(speed = c(5, 10, 20)))

}
