% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIBench_experiment.R
\name{MIBench_experiment}
\alias{MIBench_experiment}
\title{Run an experiment with MIBench}
\usage{
MIBench_experiment(
  dgp = NULL,
  MIalgorithm = NULL,
  m = 10,
  store_runs = FALSE,
  suffix = 1,
  ...
)
}
\arguments{
\item{dgp}{A dgp function in the format of MIBench}

\item{MIalgorithm}{A multiple imputation algorithm in the format of MIBench}

\item{m}{The number of imputations}

\item{store_runs}{Set to TRUE if you want to store the output of the experiments. Default is FALSE.}

\item{suffix}{A suffix for the files, especially useful if you run multiple experiments in a loop.}

\item{...}{Additional arguments for the dgp function}
}
\value{
A list of class "MIbench_imputations"
}
\description{
Run an experiment with MIBench
}
\examples{
df <- cars
regression <- lm(log(dist)~speed, data = df)
# Specifiying no scenario to simulate at the mean of speed.
simloglm(regression)
# Explicitily specifying a scenario.
simloglm(regression, scenario = list(speed = c(5, 10, 20)))

}
