% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repeat_MIBench_experiment.R
\name{repeat_MIbench_experiment}
\alias{repeat_MIbench_experiment}
\title{Repeat a MIBench_experiment multiple times}
\usage{
repeat_MIbench_experiment(
  dgp,
  MIalgorithm,
  m = 10,
  store_runs = TRUE,
  n_repetitions = 1000,
  n_cores = 4,
  ...
)
}
\arguments{
\item{dgp}{A dgp function in the format of MIBench}

\item{MIalgorithm}{A multiple imputation algorithm in the format of MIBench}

\item{m}{The number of imputations}

\item{store_runs}{Set to TRUE if you want to store the output of the experiments. Default is FALSE.}

\item{n_repetitions}{The number of repetitions (default is 1000) with fresh draws from the dgp function.}

\item{n_cores}{The number of cores for parallel processing of the experiments.}

\item{...}{Additional arguments for the dgp function}
}
\value{
A list of class "MIbench_imputations"
}
\description{
Repeat a MIBench_experiment multiple times
}
\examples{
df <- cars
regression <- lm(log(dist)~speed, data = df)
# Specifiying no scenario to simulate at the mean of speed.
simloglm(regression)
# Explicitily specifying a scenario.
simloglm(regression, scenario = list(speed = c(5, 10, 20)))
}
