% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_MIBench_all_dgps.R
\name{run_MIBench_all_dgps}
\alias{run_MIBench_all_dgps}
\title{Run all experiments for one multiple imputation algorithm}
\usage{
run_MIBench_all_dgps(
  MIalgorithm,
  m = 10,
  store_runs = FALSE,
  store_results = FALSE,
  n_repetitions = 1000,
  n_cores = 4,
  seed = NULL
)
}
\arguments{
\item{MIalgorithm}{A multiple imputation algorithm in the format of MIBench}

\item{m}{The number of imputations}

\item{store_runs}{Set to TRUE if you want to store the imputations and all data of the experiments on disk. Default is FALSE.}

\item{store_results}{Set to TRUE if you want to store the summary of the results of the experiments on disk. Default is FALSE.}

\item{n_repetitions}{The number of repetitions (default is 1000) with fresh draws from the dgp function.}

\item{n_cores}{The number of cores for parallel processing of the experiments.}

\item{seed}{A random seed for the experiments. Note that a exact replication depends on the seed and the number of cores \code{n_cores}.}
}
\value{
A list of class "MIbench_imputations"
}
\description{
Run all experiments for one multiple imputation algorithm
}
\examples{
df <- cars
regression <- lm(log(dist)~speed, data = df)
# Specifiying no scenario to simulate at the mean of speed.
simloglm(regression)
# Explicitily specifying a scenario.
simloglm(regression, scenario = list(speed = c(5, 10, 20)))
}
